Remove some sections from the postgres style guide
These sections either don't apply as we generally don't write full SQL queries,
or completely contradict rails conventions.

I think it's good for us to have some conventions around SQL, since there are
places where it makes sense to drop in to SQL for either part of a query using
ActiveRecord or for a migration, but it doesn't make sense to name tables
against the rails convensions.